[REFACTOR] guards
